server/requirements.txt
fastapi==0.109.0
uvicorn[standard]
python-dotenv==1.0.0
pydantic==2.6.4
pydantic-settings==2.2.1
sentence-transformers
numpy==1.26.0
tavily-python==0.3.5
trafilatura==1.6.1
google-generativeai==0.3.2
google-genai
lxml_html_clean